--- src/components/CollectionDetailPage/CollectionAction/index.ts
@@ -1,3 +1,0 @@
-import CollectionAction from './CollectionAction.container'
-
-export default CollectionAction
