+++ src/components/CollectionDetailPage/CollectionContextMenu/index.ts
@@ -1,0 +1,2 @@
+import CollectionContextMenu from './CollectionContextMenu.container'
+export default CollectionContextMenu
