--- src/components/CollectionDetailPage/CollectionMenu/CollectionMenu.css
@@ -1,24 +1,0 @@
-.CollectionMenu.ui.dropdown {
-  border: 1px solid var(--gray);
-  border-radius: 5px;
-  border-top-left-radius: 0;
-  border-bottom-left-radius: 0;
-  border-left: 0;
-}
-
-.CollectionMenu.ui.dropdown .icon.ellipsis {
-  margin: auto !important;
-}
-
-.CollectionMenu.ui.dropdown .item a,
-.CollectionMenu.ui.dropdown .item a:hover {
-  color: var(--text);
-}
-
-.CollectionMenu.ui.dropdown .menu > .disabled.item {
-  pointer-events: all;
-}
-
-.CollectionMenu.ui.dropdown .menu > .disabled.item a {
-  cursor: default;
-}
