--- src/components/CollectionDetailPage/CollectionMenu/index.ts
@@ -1,2 +1,0 @@
-import CollectionMenu from './CollectionMenu.container'
-export default CollectionMenu
