+++ src/components/CollectionDetailPage/CollectionPublishButton/CollectionPublishButton.tsx
@@ -10,10 +10,11 @@ import { ContractName, getContract } from 'decentraland-transactions'
 import { AuthorizationModal } from 'components/AuthorizationModal'
 import { isComplete } from 'modules/item/utils'
 import { SyncStatus } from 'modules/item/types'
-import { Props } from './CollectionAction.types'
+import { Props } from './CollectionPublishButton.types'
 import UnderReview from './UnderReview'
 
-const CollectionAction = ({ wallet, collection, items, authorizations, status, hasPendingCuration, onPublish, onPush, onInit }: Props) => {
+const CollectionPublishButton = (props: Props) => {
+  const { wallet, collection, items, authorizations, status, hasPendingCuration, onPublish, onPush, onInit } = props
   const [isAuthModalOpen, setIsAuthModalOpen] = useState(false)
 
   useEffect(() => {
@@ -92,4 +93,4 @@ const CollectionAction = ({ wallet, collection, items, authorizations, status, h
   )
 }
 
-export default React.memo(CollectionAction)
+export default React.memo(CollectionPublishButton)
