+++ src/components/CollectionDetailPage/CollectionPublishButton/CollectionPublishButton.types.ts
