+++ src/components/ThirdPartyCollectionDetailPage/CollectionContextMenu/CollectionContextMenu.css
@@ -1,0 +1,21 @@
+.CollectionContextMenu.ui.dropdown {
+  border: 1px solid var(--gray);
+  border-radius: 5px;
+}
+
+.CollectionContextMenu.ui.dropdown .icon.ellipsis {
+  margin: auto !important;
+}
+
+.CollectionContextMenu.ui.dropdown .item a,
+.CollectionContextMenu.ui.dropdown .item a:hover {
+  color: var(--text);
+}
+
+.CollectionContextMenu.ui.dropdown .menu > .disabled.item {
+  pointer-events: all;
+}
+
+.CollectionContextMenu.ui.dropdown .menu > .disabled.item a {
+  cursor: default;
+}
