--- src/components/ThirdPartyCollectionDetailPage/CollectionMenu/CollectionMenu.css
@@ -1,21 +1,0 @@
-.CollectionMenu.ui.dropdown {
-  border: 1px solid var(--gray);
-  border-radius: 5px;
-}
-
-.CollectionMenu.ui.dropdown .icon.ellipsis {
-  margin: auto !important;
-}
-
-.CollectionMenu.ui.dropdown .item a,
-.CollectionMenu.ui.dropdown .item a:hover {
-  color: var(--text);
-}
-
-.CollectionMenu.ui.dropdown .menu > .disabled.item {
-  pointer-events: all;
-}
-
-.CollectionMenu.ui.dropdown .menu > .disabled.item a {
-  cursor: default;
-}
