+++ src/components/ThirdPartyCollectionDetailPage/CollectionPublishButton/CollectionPublishButton.types.ts
