+++ src/components/ThirdPartyCollectionDetailPage/CollectionPublishButton/index.ts
@@ -1,0 +1,3 @@
+import CollectionPublishButton from './CollectionPublishButton.container'
+
+export default CollectionPublishButton
